Add CASCADE to DROP statements
This is necessary while viewr doesn’t keep track of managed views.

Currently changing a view’s dependencies might result in a dependency
tree that differs from when views were created in the first place. In
that case it may not be possible to 'cleanly' drop all views’
dependencies before dropping the view itself.